requirements.txt
agate==1.9.1
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
babel==2.17.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
daff==1.4.2
db-dtypes==1.5.0
dbt-adapters==1.22.5
dbt-bigquery==1.11.0
dbt-common==1.37.2
dbt-core==1.11.2
dbt-extractor==0.6.0
dbt-protos==1.0.422
dbt-semantic-interfaces==0.9.0
debugpy==1.8.19
decorator==5.2.1
deepdiff==8.6.1
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
executing==2.2.1
fastjsonschema==2.21.2
fonttools==4.61.1
google-api-core==2.29.0
google-auth==2.47.0
google-auth-oauthlib==1.2.4
google-cloud-aiplatform==1.134.0
google-cloud-bigquery==3.40.0
google-cloud-core==2.5.0
google-cloud-dataproc==5.24.0
google-cloud-resource-manager==1.16.0
google-cloud-storage==3.1.1
google-crc32c==1.8.0
google-genai==1.60.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
leather==0.4.1
MarkupSafe==3.0.3
mashumaro==3.14
matplotlib==3.10.8
matplotlib-inline==0.2.1
more-itertools==10.8.0
msgpack==1.1.2
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.1
oauthlib==3.3.1
orderly-set==5.5.0
packaging==25.0
pandas==2.3.3
pandas-gbq==0.33.0
parsedatetime==2.6
parso==0.8.5
pathspec==0.12.1
pillow==12.1.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==6.33.4
psutil==7.2.1
psycopg2==2.9.11
pure_eval==0.2.3
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydata-google-auth==1.9.1
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
rsa==4.9.1
scipy==1.17.0
seaborn==0.13.2
setuptools==80.10.1
six==1.17.0
sniffio==1.3.1
snowplow-tracker==1.1.0
SQLAlchemy==2.0.45
sqlparse==0.5.4
stack-data==0.6.3
tenacity==9.1.2
text-unidecode==1.3
tornado==6.5.4
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.2.14
websockets==15.0.1
zipp==3.23.0
